Fix up installer tests
