+++ Installer/.gitignore
@@ -1,4 +1,3 @@
 dist
 *.tgz
 UseThisWireGuardConfigurationToConnectToYourWirtBot.conf
-!tests/**/wirtbot-installer.config.json
