+++ build-automation/Installer-Tests/.gitignore
@@ -1,0 +1,1 @@
+!tests/**/wirtbot-installer.config.json
