+++ build-automation/Installer-Tests/tests/test-install-with-ssh-password/wirtbot-installer.config.json
