+++ build-automation/Installer-Tests/tests/test-update/wirtbot-installer.config.json
